Chopped out the reverse data flow
