+++ zellij-tile/src/lib.rs
@@ -1,13 +1,16 @@
 mod shim;
+mod data;
 
 pub use shim::*;
+pub use data::*;
 #[allow(unused_variables)]
 pub trait ZellijTile {
-    fn init(&mut self) {}
+    fn load(&mut self) {}
+    fn update(&mut self, dt: f64) {}
+    // FIXME: I might want to make `draw()` immutable with just `&self`
     fn draw(&mut self, rows: usize, cols: usize) {}
     fn handle_key(&mut self, key: Key) {}
     fn handle_global_key(&mut self, key: Key) {}
-    fn update_tabs(&mut self, active_tab_index: usize, num_active_tabs: usize) {}
 }
 
 #[macro_export]
@@ -19,7 +22,7 @@ macro_rules! register_tile {
 
         fn main() {
             STATE.with(|state| {
-                state.borrow_mut().init();
+                state.borrow_mut().load();
             });
         }
 
@@ -42,15 +45,6 @@ macro_rules! register_tile {
             STATE.with(|state| {
                 state.borrow_mut().handle_global_key($crate::get_key());
             });
-        }
-
-        #[no_mangle]
-        pub fn update_tabs(active_tab_index: i32, num_active_tabs: i32) {
-            STATE.with(|state| {
-                state
-                    .borrow_mut()
-                    .update_tabs(active_tab_index as usize, num_active_tabs as usize);
-            })
         }
     };
 }
